# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Dokumentasi API untuk CRUD Admin WA Rotator",
        "version": "1.0"
    },
    "paths": {
        "/api/category-location-admin": {
            "get": {
                "tags": [
                    "Category Locations"
                ],
                "summary": "Menampilkan daftar Lokasi",
                "description": "Mengambil daftar Lokasi dengan fitur pencarian berdasarkan name atau deskripsi",
                "operationId": "8e9f39379f391dc2b8700a77b152fa03",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Filter Lokasi berdasarkan name atau deskripsi",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Malang"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Users retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Users retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "john_doe"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "........."
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/category-location-create": {
            "post": {
                "tags": [
                    "Category Locations"
                ],
                "summary": "Create a new category location",
                "operationId": "d8bfbf64c43ff1635f92c5d5a6e0f88a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "description": {
                                        "type": "string",
                                        "maxLength": 255
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Category location created successfully"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            }
        },
        "/api/category-location-update/{id}": {
            "put": {
                "tags": [
                    "Category Locations"
                ],
                "summary": "Update an existing category location",
                "operationId": "68beef3b6c5ed2554dbccad6fc734c4f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the category location",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "description": {
                                        "type": "string",
                                        "maxLength": 255
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Category location updated successfully"
                    },
                    "404": {
                        "description": "Category location not found"
                    }
                }
            }
        },
        "/api/category-location-delete/{id}": {
            "delete": {
                "tags": [
                    "Category Locations"
                ],
                "summary": "Delete a category location",
                "operationId": "24ed08ce839417f4c7bc09445035deb9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the category location",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category location deleted successfully"
                    },
                    "404": {
                        "description": "Category location not found"
                    }
                }
            }
        },
...
    },
...
}